Added test for Downloader.

+++ DevTools/Downloader/src/main/java/info/smart_tools/smartactors/downloader/files/actors/FileOperationsActor.java
@@ -26,7 +26,7 @@ public class FileOperationsActor extends StatelessActor {
     @Executable()
     public Path copyFile(final Path source, final Path target) {
         try {
-            if (source.toFile().exists() && null != target) {
+            if (null != source && source.toFile().exists() && null != target) {
                 Files.copy(source, target);
             }
